Disregard gestures that registered a physical press
If we received a physical button press (BTN_LEFT, BTN_RIGHT) while
handling a gesture, that means that the user either pressed hard enough
to trigger a "force touch" event or pressed the physical button during
the gesture.

In all cases, ignore the gesture and do not take any action until the
start of a completely new gesture is detected.