Use configuration-dependent logging macros
This makes trace/debug/info/warn macros a no-op when compiled without
the logging feature enabled (like using the `log/max_level_error`
feature but without requiring `--default-features false` in order to use
`--features logging`).

The error!() macro is converted to an eprintln!() call when logging is
disabled so we can now use error!() instead of eprintln!() to get nicer
error messages when running with logging enabled without risking their
disappearing when logging is disabled (or needing to ship with a
dependency on (pretty_)env_logger for regular builds).