Compensate for drivers that don't produce ABS_X/ABS_Y
The Apple Magic Mouse does not seem to emit any "overall position" ABS
events (ABS_X and ABS_Y). It only generates the per-tool
ABS_MT_POSITION_X and ABS_MT_POSITION_Y events.

In this case, we just pick one of the tools to be a stand-in for our
overall position and use that as the location. It should be the same
tool each time since we only do this while the maximum finger count (for
the current gesture) is reached.