Do not always revert to slot_id 0 after SYN_REPORT
It's possible for some drivers (like the Magic Mouse one) to assume that
the ABS_MT_SLOT (and its associated ABS_MT_TRACKING_ID) are persisted
across SYN_REPORT events.

We store the ABS_MT_SLOT value when we come across it and use that to
initialize `slot` in `TouchpadState::process()`, only resetting it after
getting an explicit ABS_MT_TRACKING_ID of -1 (which we should interpret
as freeing the binding between the slot id and the tool's tracking id).