Correct README notes on adding a user to the `input` group
While logging out of a tty session then back in is sufficient for the group
privileges to be re-read, it does not seem that logging out of the login manager
and back in is sufficient. Since a touchpad will be used almost exclusively in
GUI environments, change the guidance to just say that a reboot is required.