Add another option for emulating middle click to the sample config
`xdotool click 2` doesn't appear to work because something in the input stack
seems to just ignore middle clicks when the device is a touchpad?

The new workaround added to the sample configuration file works great, but is
primarily for use in web browsers. It might work in other apps too, but it's not
a "native" middle click - it takes advantage of the workarounds web browsers
have added to work around lack of middle click for laptop/touchpad users.